Update info.plist for Mechanic
Sorry for the seriously late response about this, but this is the correct key for `info.plist`. I think that you had this set up before correctly, but the listing on the server was set with the incorrect `repository` key. I've fix this on the server and everything is working correctly now.

Closes #1